+++ HerkesYazarOlsun.Model/Entity/BooksComment.cs
@@ -8,7 +8,7 @@ namespace HerkesYazarOlsun.Model.Entity
         public string NAME { get; set; }
         public string EMAIL { get; set; }
         public string ACIKLAMA { get; set; }
-        public int? BookId { get; set; }
+        public long? BookId { get; set; }
         public Books? Books { get; set; }
         public long LoginUserId { get; set; }
         public int? StarPuani { get; set; }
