+++ HerkesYazarOlsun.Model/Entity/BooksDegerlendirme.cs
@@ -9,7 +9,7 @@ namespace HerkesYazarOlsun.Model.Entity
         public string EMAIL { get; set; }
         public string KONU { get; set; }
         public string ACIKLAMA { get; set; }
-        public int? BookId { get; set; }
+        public long? BookId { get; set; }
         public Books? Books { get; set; }
         public long LoginUserId { get; set; }       
         public int? StarPuani { get; set; }
